requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.5.2
    # via
    #   channels
    #   channels-redis
    #   django
    #   djoser
async-timeout==4.0.2
    # via redis
certifi==2022.9.24
    # via requests
cffi==1.15.1
    # via cryptography
channels==4.0.0
    # via
    #   -r requirements.in
    #   channels-redis
channels-redis==4.0.0
    # via -r requirements.in
charset-normalizer==2.1.1
    # via requests
coreapi==2.3.3
    # via
    #   djoser
    #   drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
cryptography==38.0.1
    # via social-auth-core
defusedxml==0.7.1
    # via
    #   python3-openid
    #   social-auth-core
deprecated==1.2.13
    # via redis
django==4.1.2
    # via
    #   -r requirements.in
    #   channels
    #   django-debug-toolbar
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-nested-routers
    #   drf-yasg
django-cleanup==6.0.0
    # via -r requirements.in
django-debug-toolbar==3.7.0
    # via -r requirements.in
django-decouple==2.1
    # via -r requirements.in
django-filter==22.1
    # via -r requirements.in
django-templated-mail==1.1.1
    # via djoser
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-nested-routers
    #   drf-yasg
djangorestframework-simplejwt==4.8.0
    # via
    #   -r requirements.in
    #   djoser
djoser==2.1.0
    # via -r requirements.in
drf-nested-routers==0.93.4
    # via -r requirements.in
drf-yasg==1.21.4
    # via -r requirements.in
idna==3.4
    # via requests
inflection==0.5.1
    # via drf-yasg
itypes==1.2.0
    # via coreapi
jinja2==3.1.2
    # via coreschema
markupsafe==2.1.1
    # via jinja2
msgpack==1.0.4
    # via channels-redis
oauthlib==3.2.2
    # via
    #   requests-oauthlib
    #   social-auth-core
packaging==21.3
    # via
    #   drf-yasg
    #   redis
pillow==9.2.0
    # via -r requirements.in
psycopg2-binary==2.9.4
    # via -r requirements.in
pycparser==2.21
    # via cffi
pyjwt==2.6.0
    # via
    #   djangorestframework-simplejwt
    #   social-auth-core
pyotp==2.7.0
    # via -r requirements.in
pyparsing==3.0.9
    # via packaging
python3-openid==3.2.0
    # via social-auth-core
pytz==2022.5
    # via
    #   djangorestframework
    #   drf-yasg
redis==4.3.4
    # via channels-redis
requests==2.28.1
    # via
    #   -r requirements.in
    #   coreapi
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==1.3.1
    # via social-auth-core
ruamel-yaml==0.17.21
    # via drf-yasg
ruamel-yaml-clib==0.2.7
    # via ruamel-yaml
six==1.16.0
    # via social-auth-app-django
social-auth-app-django==4.0.0
    # via djoser
social-auth-core==4.3.0
    # via social-auth-app-django
sqlparse==0.4.3
    # via
    #   django
    #   django-debug-toolbar
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.12
    # via requests
wrapt==1.14.1
    # via deprecated
